refactor(core): centralizar validaciones y guards para mejorar la modularidad
Mueve los archivos de validación de los módulos Auth, User y Address a una carpeta utils central.

Crea un GuardsModule para reutilizar los guards en toda la aplicación.

Elimina las dependencias directas en los controladores.
