+++ src/modules/auth/auth.module.ts
@@ -3,15 +3,13 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { AuthController } from './controllers/auth.controller';
 import { AuthService } from './services/auth.service';
 import { HttpModule } from '@nestjs/axios';
-import { RolesGuard } from './guards/roles.guard';
-import { JwtAuthGuard } from './jwt/jwt.guard';
 import { JwtModule } from './jwt/jwt.module';
-
+import { GuardsModule } from 'src/guards/guards.module';
 
 @Module({
-  imports:[HttpModule,JwtModule],
+  imports: [HttpModule, JwtModule, GuardsModule],
   controllers: [AuthController],
-  providers: [AuthService, PrismaService,JwtAuthGuard,RolesGuard],
-  exports: [JwtAuthGuard, RolesGuard,JwtModule],
+  providers: [AuthService, PrismaService],
+  exports: [JwtModule, GuardsModule],
 })
 export class AuthModule {}
